backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pymysql
python-dotenv
PyJWT
cryptography
bcrypt
email-validator
password-validator